+++ e2e/wallets/intergration-coins/src/index.ts
@@ -84,14 +84,14 @@ const test_service = async function (this: any) {
             // 'AVAX',  //fast
             // 'BASE',  //fast
             // 'BSC',   //fast
-            'BTC',
+            // 'BTC',
             // 'BCH',
             // 'GAIA',
             // 'OSMO',
             // 'XRP',
             // 'DOGE',
             // 'DASH',
-            // 'MAYA',
+            'MAYA',
             // 'LTC',
             // 'THOR'
         ]
@@ -137,104 +137,104 @@ const test_service = async function (this: any) {
 
         //add paths to keepkey
         //add account 0 p2sh segwit
-        paths.push({
-            note: 'Bitcoin account 0 segwit (p2sh)',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2sh-p2wpkh',
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'xpub',
-            addressNList: [0x80000000 + 49, 0x80000000 + 0, 0x80000000 + 0],
-            addressNListMaster: [0x80000000 + 49, 0x80000000 + 0, 0x80000000 + 0, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
-
-        //add account1
-        paths.push({
-            note: 'Bitcoin account 0 Native Segwit (Bech32)',
-            blockchain: 'bitcoin',
-            symbol: 'BTC',
-            symbolSwapKit: 'BTC',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2wpkh', //bech32
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'zpub',
-            addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
-            addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
-
-        paths.push({
-            note: 'Bitcoin account 1 legacy',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2pkh',
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'xpub',
-            addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 1],
-            addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
-
-        //add account1
-        paths.push({
-            note: 'Bitcoin account 1 Native Segwit (Bech32)',
-            blockchain: 'bitcoin',
-            symbol: 'BTC',
-            symbolSwapKit: 'BTC',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2wpkh', //bech32
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'zpub',
-            addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
-            addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
-
-        paths.push({
-            note: 'Bitcoin account 1 legacy',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2pkh',
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'xpub',
-            addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 2],
-            addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 2, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
-
-        //add account3
-        paths.push({
-            note: 'Bitcoin account 1 Native Segwit (Bech32)',
-            blockchain: 'bitcoin',
-            symbol: 'BTC',
-            symbolSwapKit: 'BTC',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2wpkh', //bech32
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'zpub',
-            addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
-            addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
-
-        paths.push({
-            note: 'Bitcoin account 3 legacy',
-            blockchain: 'bitcoin',
-            symbol: 'BTC',
-            symbolSwapKit: 'BTC',
-            networks: ['bip122:000000000019d6689c085ae165831e93'],
-            script_type: 'p2pkh',
-            available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
-            type: 'xpub',
-            addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 3],
-            addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 3, 0, 0],
-            curve: 'secp256k1',
-            showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
-        });
+        // paths.push({
+        //     note: 'Bitcoin account 0 segwit (p2sh)',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2sh-p2wpkh',
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'xpub',
+        //     addressNList: [0x80000000 + 49, 0x80000000 + 0, 0x80000000 + 0],
+        //     addressNListMaster: [0x80000000 + 49, 0x80000000 + 0, 0x80000000 + 0, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
+        //
+        // //add account1
+        // paths.push({
+        //     note: 'Bitcoin account 0 Native Segwit (Bech32)',
+        //     blockchain: 'bitcoin',
+        //     symbol: 'BTC',
+        //     symbolSwapKit: 'BTC',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2wpkh', //bech32
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'zpub',
+        //     addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
+        //     addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
+        //
+        // paths.push({
+        //     note: 'Bitcoin account 1 legacy',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2pkh',
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'xpub',
+        //     addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 1],
+        //     addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
+        //
+        // //add account1
+        // paths.push({
+        //     note: 'Bitcoin account 1 Native Segwit (Bech32)',
+        //     blockchain: 'bitcoin',
+        //     symbol: 'BTC',
+        //     symbolSwapKit: 'BTC',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2wpkh', //bech32
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'zpub',
+        //     addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
+        //     addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
+        //
+        // paths.push({
+        //     note: 'Bitcoin account 1 legacy',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2pkh',
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'xpub',
+        //     addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 2],
+        //     addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 2, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
+        //
+        // //add account3
+        // paths.push({
+        //     note: 'Bitcoin account 1 Native Segwit (Bech32)',
+        //     blockchain: 'bitcoin',
+        //     symbol: 'BTC',
+        //     symbolSwapKit: 'BTC',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2wpkh', //bech32
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'zpub',
+        //     addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
+        //     addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
+        //
+        // paths.push({
+        //     note: 'Bitcoin account 3 legacy',
+        //     blockchain: 'bitcoin',
+        //     symbol: 'BTC',
+        //     symbolSwapKit: 'BTC',
+        //     networks: ['bip122:000000000019d6689c085ae165831e93'],
+        //     script_type: 'p2pkh',
+        //     available_scripts_types: ['p2pkh', 'p2sh', 'p2wpkh', 'p2sh-p2wpkh'],
+        //     type: 'xpub',
+        //     addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 3],
+        //     addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 3, 0, 0],
+        //     curve: 'secp256k1',
+        //     showDisplay: false, // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        // });
 
 
         // if(blockchains.includes('bip122:000000000019d6689c085ae165831e93')){
@@ -308,8 +308,6 @@ const test_service = async function (this: any) {
         // await app.getGasAssets()
         // await app.getPubkeys()
         // await app.getBalances()
-
-
         //clear cache
 
         // log.info(tag,"resultInit: ",resultInit)
@@ -328,7 +326,6 @@ const test_service = async function (this: any) {
         }
         log.info(tag,' ****** Validated Assets for caips ******')
 
-
         let pubkeys
         if(app.pubkeys.length === 0){
             log.info(tag,'cache is empty refreshing... ')
@@ -337,7 +334,6 @@ const test_service = async function (this: any) {
             pubkeys = app.pubkeys
             log.info(tag,'cache found! ', pubkeys.length)
         }
-
 
         log.info(tag,"pubkeys: ",pubkeys.length)
         assert(pubkeys)