+++ e2e/swaps/e2e-swap-suite/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @coinmasters/e2e-transfer-dogecoin
 
+## 1.8.26
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.26
+
+## 1.8.25
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.25
+
+## 1.8.24
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.24
+
+## 1.8.23
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.23
+
+## 1.8.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.22
+
 ## 1.8.21
 
 ### Patch Changes