+++ packages/pioneer/pioneer-react/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @coinmasters/pioneer-react
 
+## 0.8.26
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.26
+
+## 0.8.25
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.25
+
+## 0.8.24
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.24
+
+## 0.8.23
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.23
+
+## 0.8.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.22
+
 ## 0.8.21
 
 ### Patch Changes