+++ packages/pioneer/pioneer-sdk/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @coinmasters/pioneer-sdk
 
+## 4.8.26
+
+### Patch Changes
+
+- bump
+
+## 4.8.25
+
+### Patch Changes
+
+- bump
+
+## 4.8.24
+
+### Patch Changes
+
+- bump
+
+## 4.8.23
+
+### Patch Changes
+
+- bump
+
+## 4.8.22
+
+### Patch Changes
+
+- bump
+
 ## 4.8.21
 
 ### Patch Changes